+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## master
 
+- [#98]: add a validation error to the current form if a form of a previous step became invalid
 - BC breaks (follow `UPGRADE-3.0.md` to upgrade):
   - [#101]: support for concurrent instances of the same flow
   - removed the step field template
 
+[#98]: https://github.com/craue/CraueFormFlowBundle/issues/98
 [#101]: https://github.com/craue/CraueFormFlowBundle/issues/101
 
 ## 2.1.4 (2013-12-05)